PLUGINS - ensure wavesurfer has not been destroyed when unmounting plugins components

+++ src/plugins/minimap.tsx
@@ -57,7 +57,9 @@ export default class Minimap extends PureComponent<IMinimapProps, {}> {
    * @memberof Minimap
    */
   componentWillUnmount(): void {
-    this.props.wavesurfer.destroyPlugin('minimap');
+    if (!this.props.wavesurfer.isDestroyed) {
+      this.props.wavesurfer.destroyPlugin('minimap');
+    }
   }
 
   /**
